refactor: move recommend-issues to bot directory and rename to bot-recommend-issues
